Split autopilot.cpp in header and source files
